# api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted media server for movies, TV shows, and music",
        "title": "ViewRA Media Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ViewRA Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/libraries": {
            "get": {
                "description": "Returns a list of all libraries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "libraries"
                ],
                "summary": "List all libraries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.ListLibrariesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new media library",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "libraries"
                ],
                "summary": "Create a new library",
                "parameters": [
                    {
                        "description": "Library details",
                        "name": "library",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.CreateLibraryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.LibraryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/libraries/{id}": {
            "get": {
                "description": "Returns details of a specific library",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "libraries"
                ],
                "summary": "Get a library by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.GetLibraryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing library",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "libraries"
                ],
                "summary": "Update a library",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Library details",
                        "name": "library",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.UpdateLibraryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.LibraryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a library and all its media entries (does not delete files)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "libraries"
                ],
                "summary": "Delete a library",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/libraries/{id}/scan": {
            "post": {
                "description": "Scans a library for media files",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "libraries"
                ],
                "summary": "Scan a library",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.StartScanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Library is already being scanned",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_viewra_viewra_internal_application_library.CreateLibraryRequest": {
            "type": "object",
            "required": [
                "name",
                "path",
                "type"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "movies",
                        "tv",
                        "music"
                    ]
                }
            }
        },
        "github_com_viewra_viewra_internal_application_library.GetLibraryResponse": {
            "type": "object",
            "properties": {
                "library": {
                    "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.LibraryResponse"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_library.LibraryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_library.ListLibrariesResponse": {
            "type": "object",
            "properties": {
                "libraries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_viewra_viewra_internal_application_library.LibraryResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_library.StartScanResponse": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "integer"
                },
                "library_id": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_library.UpdateLibraryRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "movies",
                        "tv",
                        "music"
                    ]
                }
            }
        },
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}